Revert "split less files"
This reverts commit 8e11501f1a59d853ce4912dc10d53810e45c414b.
